process: rationalize the way a process's cwd is inherited
Add a PROCESS_INHERITS_CWD flag that one can pass to exec() in order to ask for the current process'
working directory to be inherited by the spawned process. This requires changing exec() from taking
a "bool fg" to taking a "uint32_t flags" which includes PROCESS_IS_FOREGROUND (or not) + the new
working directory inheritance flag.

Convert exec() to using argtypes over argc since we're changing things anyway. Doing that requires changes
in libuserspace as well as newlib.

This is technically done in a backwards compatible manner as 1 stands to mean "foreground" and 0 "background", same
as before the change - and PROCESS_INHERITS_CWD == 2, but previous builds did not have a notion of working directory,
so this is without loss of compatibility.

More work towards https://github.com/egranata/puppy/issues/23

+++ newlib/syscalls/interface.cpp
@@ -308,7 +308,7 @@ NEWLIB_IMPL_REQUIREMENT int ioctl(int fd, int a, int b) {
 }
 
 NEWLIB_IMPL_REQUIREMENT int spawn(const char* path, const char* args, int flags) {
-    auto eo = exec_syscall((uint32_t)path, (uint32_t)args, flags);
+    auto eo = exec_syscall(path, args, flags);
     if (eo & 1) return -1;
     return eo >> 1;
 }
